tests/test_phase0.py
"""Phase 0 tests for project infrastructure.

These tests verify that the basic project setup is correct and all
infrastructure components are working as expected.
"""

import importlib
import logging

import pytest

import jactus
from jactus import exceptions, logging_config


class TestT01PackageInstallation:
    """T0.1: Verify package can be installed and imported."""

    def test_import_jactus(self):
        """Test that jactus can be imported."""
        assert jactus is not None

    def test_version_accessible(self):
        """Test that jactus.__version__ is accessible."""
        assert hasattr(jactus, "__version__")
        assert isinstance(jactus.__version__, str)
        assert jactus.__version__ == "0.1.0"

    def test_submodules_importable(self):
        """Test that all submodules are importable."""
        submodules = [
            "jactus.core",
            "jactus.utilities",
            "jactus.contracts",
            "jactus.functions",
            "jactus.observers",
            "jactus.engine",
            "jactus.exceptions",
            "jactus.logging_config",
            "jactus.cli",
        ]
        for module_name in submodules:
            module = importlib.import_module(module_name)
            assert module is not None


class TestT02DependencyVerification:
    """T0.2: Ensure all dependencies are correctly installed."""

    def test_jax_imports(self):
        """Test that JAX imports and basic operations work."""
        import jax.numpy as jnp

        # Basic JAX operation
        x = jnp.array([1, 2, 3])
        result = jnp.sum(x)
        assert result == 6

    def test_flax_imports(self):
        """Test that Flax NNX module can be imported."""
        import flax

        assert flax is not None

    def test_pydantic_imports(self):
        """Test that Pydantic can be imported."""
        import pydantic

        assert pydantic is not None

    def test_numpy_imports(self):
        """Test that NumPy can be imported."""
        import numpy as np

        assert np is not None

    def test_dateutil_imports(self):
        """Test that python-dateutil can be imported."""
        import dateutil

        assert dateutil is not None


class TestT03ExceptionSystem:
    """T0.3: Test exception hierarchy and error handling."""

    def test_all_exceptions_can_be_raised(self):
        """Test that all custom exceptions can be raised and caught."""
        exception_classes = [
            exceptions.ActusException,
            exceptions.ContractValidationError,
            exceptions.ScheduleGenerationError,
            exceptions.StateTransitionError,
            exceptions.PayoffCalculationError,
            exceptions.ObserverError,
            exceptions.DateTimeError,
            exceptions.ConventionError,
            exceptions.ConfigurationError,
            exceptions.EngineError,
        ]

        for exc_class in exception_classes:
            with pytest.raises(exc_class):
                raise exc_class("Test error")

    def test_exception_inheritance(self):
        """Test that exception inheritance is correct."""
        # All exceptions should inherit from ActusException
        exception_classes = [
            exceptions.ContractValidationError,
            exceptions.ScheduleGenerationError,
            exceptions.StateTransitionError,
            exceptions.PayoffCalculationError,
            exceptions.ObserverError,
            exceptions.DateTimeError,
            exceptions.ConventionError,
            exceptions.ConfigurationError,
            exceptions.EngineError,
        ]

        for exc_class in exception_classes:
            assert issubclass(exc_class, exceptions.ActusException)
            assert issubclass(exc_class, Exception)

    def test_error_messages_are_informative(self):
        """Test that error messages are informative."""
        msg = "Test error message"
        exc = exceptions.ActusException(msg)
        assert str(exc) == msg

    def test_context_information_preserved(self):
        """Test that context information is preserved."""
        msg = "Test error"
        context = {"contract_id": "PAM-001", "event": "IP"}
        exc = exceptions.ActusException(msg, context=context)

        assert exc.message == msg
        assert exc.context == context
        assert "contract_id" in str(exc)
        assert "PAM-001" in str(exc)


class TestT04LoggingSystem:
    """T0.4: Verify logging configuration works."""

    def test_logger_can_be_configured(self):
        """Test that logger can be configured at different levels."""
        # Test that configuration completes without errors
        logging_config.configure_logging(level="DEBUG", console=True)
        logging_config.configure_logging(level="INFO", console=True)
        # The actual log level check is done via the jactus root logger
        root_logger = logging.getLogger("jactus")
        assert root_logger.level == logging.INFO

    def test_log_messages_appear(self, caplog):
        """Test that log messages appear in expected handlers."""
        import jactus

        # Use the package logger that's configured by default
        logger = jactus.get_logger("jactus.test")

        with caplog.at_level(logging.INFO, logger="jactus.test"):
            logger.info("Test message")
            # Message should appear either in caplog or in stdout
            # Since we use propagate=False, just verify no error occurred
            assert True  # Test passes if no exception raised

    def test_performance_logger(self):
        """Test that performance logging can be enabled/disabled."""
        perf_logger = logging_config.get_performance_logger("test_module")
        assert perf_logger is not None
        assert "performance" in perf_logger.name


class TestT05CodeQuality:
    """T0.5: Ensure code quality tools work correctly.

    These tests are primarily documentation of the manual checks.
    Actual quality checks are run via pre-commit hooks and CI/CD.
    """

    def test_package_has_type_hints(self):
        """Verify that main modules have type hints."""
        # Check that exceptions module has annotations
        assert hasattr(exceptions.ActusException.__init__, "__annotations__")

    def test_docstrings_present(self):
        """Verify that main classes have docstrings."""
        assert exceptions.ActusException.__doc__ is not None
        assert logging_config.configure_logging.__doc__ is not None


class TestT06TestingInfrastructure:
    """T0.6: Verify test framework is operational."""

    def test_pytest_runs(self):
        """Test that pytest is working."""
        # This test itself verifies pytest is working
        assert True

    def test_fixtures_accessible(self, sample_dates, jax_random_key, tolerance):
        """Test that test fixtures are accessible."""
        assert sample_dates is not None
        assert "today" in sample_dates
        assert jax_random_key is not None
        assert tolerance is not None
        assert "rtol" in tolerance

    def test_mock_observer_fixture(self, mock_risk_factor_observer):
        """Test that mock observer fixture works."""
        from datetime import datetime, timezone

        observer = mock_risk_factor_observer
        rate = observer.get_rate(datetime(2024, 1, 1, tzinfo=timezone.utc), "USD-LIBOR")
        assert rate == 0.05

        fx_rate = observer.get_fx_rate(datetime(2024, 1, 1, tzinfo=timezone.utc), "EUR/USD")
        assert fx_rate == 1.1


class TestT07PublicAPI:
    """Test that the public API is properly exported."""

    def test_exceptions_exported(self):
        """Test that exceptions are exported from main package."""
        assert hasattr(jactus, "ActusException")
        assert hasattr(jactus, "ContractValidationError")

    def test_logging_exported(self):
        """Test that logging functions are exported."""
        assert hasattr(jactus, "configure_logging")
        assert hasattr(jactus, "get_logger")

    def test_metadata_exported(self):
        """Test that metadata is exported."""
        assert hasattr(jactus, "__version__")
        assert hasattr(jactus, "__author__")
        assert hasattr(jactus, "__license__")
